split out tinkergraph as a subproject
