+++ codepropertygraph/build.sbt
@@ -1,7 +1,6 @@
 name := "codepropertygraph"
 
 libraryDependencies ++= Seq(
-  "org.apache.tinkerpop" % "tinkergraph-gremlin" % "3.3.3",
   "com.michaelpollmeier" %% "gremlin-scala" % "3.3.3.15",
   "com.google.guava" % "guava" % "21.0",
   "org.apache.commons" % "commons-lang3" % "3.5",
