+++ project/Projects.scala
@@ -5,4 +5,5 @@ object Projects {
   lazy val protoBindings = project.in(file("proto-bindings"))
   lazy val queryPrimitives = project.in(file("query-primitives"))
   lazy val enhancements = project.in(file("enhancements"))
+  lazy val cpgloaderTinkergraph = project.in(file("cpgloader-tinkergraph"))
 }
