--- query-primitives/src/main/java/io/shiftleft/queryprimitives/CpgInputFormat.java
@@ -1,5 +1,0 @@
-package io.shiftleft.queryprimitives;
-
-public enum CpgInputFormat {
-  PROTOZIP, GRAPHML, GRYO, TINKERPOP
-}
