--- query-primitives/src/main/java/io/shiftleft/queryprimitives/cpgloading/tinker/TinkerFormat.java
@@ -1,5 +1,0 @@
-package io.shiftleft.queryprimitives.cpgloading.tinker;
-
-public enum TinkerFormat {
-  GRAPHML, GRYO
-}
